refactor(#4): jwt 패키지 구조 변경
