+++ src/main/java/org/example/stellog/auth/api/OAuthController.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.example.stellog.auth.api.jwt.dto.TokenDto;
 import org.example.stellog.auth.application.OAuthService;
+import org.example.stellog.global.jwt.dto.TokenDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
