+++ src/main/java/org/example/stellog/auth/api/userInfo/KakaoUserInfo.java
@@ -1,7 +1,5 @@
 package org.example.stellog.auth.api.userInfo;
 
-import org.example.stellog.auth.api.dto.UserInfo;
-
 public class KakaoUserInfo implements OAuthUserInfo {
     private final UserInfo attributes;
